requirements.txt
fastapi==0.110.1
uvicorn==0.34.3
jinja2==3.1.3
sqlmodel==0.0.24
fastapi-users==14.0.1
fastapi-users[sqlalchemy]==14.0.1
python-multipart
fastapi-mail==1.4.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
aiosqlite==0.20.0